+++ test.py
@@ -1,0 +1,4 @@
+""" some tests """
+A = 2
+B = 3
+print(f"{A + B = }")